bug fix:dc2_FS.hpp, shoud use MPI_TYPE<template Float>
I should use MPI_TYPE<template Float> but I  hardcoding MPI_TYPE<double>.

+++ src/eigen_FS/dc2_FS.hpp
@@ -106,7 +106,7 @@ void dc2_FS(int n, int nvec, Float d[], Float e[], Float z[], int ldz,
   }
 #endif
 
-  MPI_Allreduce(&eigen_dc::flops, ret, 1, FS_const::MPI_TYPE<double>, MPI_SUM,
+  MPI_Allreduce(&eigen_dc::flops, ret, 1, FS_const::MPI_TYPE<Float>, MPI_SUM,
                 eigen_comm);
 
   return;
